requirements.txt
# CropHealth AI - Python Dependencies
# Install with: pip install -r requirements.txt
# Python 3.11.9 required

# Web Framework
Flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0

# Machine Learning (Python 3.11+ compatible)
numpy==1.26.4
scikit-learn==1.5.2

# Image Processing
Pillow==10.4.0

# Additional dependencies
Werkzeug==3.0.3
